Add files via upload
Exploratory Data Analysis, EDA for short, is simply a ‘first look at the data. It forms a critical part of the machine learning workflow and it is at this stage we start to understand the data we are working with and what it contains. In essence, it allows us to make sense of the data before applying advanced analytics and machine learning.

Pandas Profiling is a Python library that allows you to generate a very detailed report on our pandas data frame without much input from the user.